# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for AppStoreConnect project",
        "title": "AppStoreConnect project API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1/{tenantId}/{integrationId}",
    "paths": {
        "/bundleId/{id}": {
            "delete": {
                "description": "BundleId deletion by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bundleId"
                ],
                "summary": "Delete bundleId",
                "operationId": "delete-bundleId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "tenantId",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "integrationId",
                        "name": "integrationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "bundleId identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deletion success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "BundleId not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/bundleIds": {
            "get": {
                "description": "Get all bundleIds",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bundleId"
                ],
                "summary": "Get bundleIds list",
                "operationId": "get-bundleIds-list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "tenantId",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "integrationId",
                        "name": "integrationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/entities.BundleId"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create new bundleId instance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bundleId"
                ],
                "summary": "Add new BundleId",
                "operationId": "add-bundleId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "tenantId",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "integrationId",
                        "name": "integrationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "bundleId params",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/d-kv_signer_api_pkg_httpserver_entities.InputCreateBundleId"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "commandID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/bundleIds/status/{id}": {
            "post": {
                "description": "Get command status by command id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bundleId"
                ],
                "summary": "Check status",
                "operationId": "get-status-bundleId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "tenantId",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "integrationId",
                        "name": "integrationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "command identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/bundleIds/{id}": {
            "get": {
                "description": "Get bundleId instance by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bundleId"
                ],
                "summary": "Get bundleId",
                "operationId": "get-bundleId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "tenantId",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "integrationId",
                        "name": "integrationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "bundleId identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.BundleId"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "d-kv_signer_api_pkg_httpserver_entities.InputCreateBundleId": {
            "type": "object",
            "properties": {
                "bundleId": {
                    "type": "string"
                },
                "bundleName": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "seedId": {
                    "type": "string"
                }
            }
        },
...
        "entities.BundleId": {
            "type": "object",
            "properties": {
                "capabilities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.CapabilityType"
                    }
                },
                "id": {
                    "type": "string"
                },
                "identifier": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "entity.CapabilityType": {
            "type": "string",
            "enum": [
                "IN_APP_PURCHASE",
                "PUSH_NOTIFICATIONS",
                "APPLE_PAY",
                "NFC_TAG_READING",
                "APPLE_ID_AUTH"
            ],
            "x-enum-varnames": [
                "InAppPurchase",
                "PushNotifications",
                "ApplePay",
                "NfcTagReading",
                "AppleIdAuth"
            ]
        },
...
        "pkg_httpserver_handlers.errorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}